+++ sdk/notificationhubs/notification-hubs/src/index.ts
@@ -2,10 +2,9 @@
 // Licensed under the MIT license.
 
 export { NotificationHubsClient } from "./notificationHubsClient.js";
-export * from "./models/device.js";
 export * from "./models/installation.js";
 export * from "./models/notification.js";
-export * from "./models/notificationBuilder.js";
+export * from "./models/notificationBodyBuilder.js";
 export * from "./models/notificationDetails.js";
 export * from "./models/notificationHubJob.js";
 export * from "./models/options.js";