+++ sdk/notificationhubs/notification-hubs/src/notificationHubsClient.ts
@@ -18,10 +18,10 @@ import {
 } from "./models/notificationDetails.js";
 import { NotificationHubJob, NotificationHubJobPoller } from "./models/notificationHubJob.js";
 import { NotificationHubsClientContext, createClientContext } from "./api/index.js";
+import { RegistrationDescription, RegistrationChannel } from "./models/registration.js";
 import { Notification } from "./models/notification.js";
 import { OperationOptions } from "@azure/core-client";
 import { PagedAsyncIterableIterator } from "@azure/core-paging";
-import { RegistrationDescription } from "./models/registration.js";
 import { beginSubmitNotificationHubJob as beginSubmitNotificationHubJobMethod } from "./api/beginSubmitNotificationHubJob.js";
 import { cancelScheduledNotification as cancelScheduledNotificationMethod } from "./api/cancelScheduledNotification.js";
 import { createOrUpdateInstallation as createOrUpdateInstallationMethod } from "./api/createOrUpdateInstallation.js";
@@ -36,7 +36,7 @@ import { getNotificationHubJob as getNotificationHubJobMethod } from "./api/getN
 import { getNotificationOutcomeDetails as getNotificationOutcomeDetailsMethod } from "./api/getNotificationOutcomeDetails.js";
 import { getRegistration as getRegistrationMethod } from "./api/getRegistration.js";
 import { listNotificationHubJobs as listNotificationHubJobsMethod } from "./api/listNotificationHubJobs.js";
-import { listRegistrationsByDevice as listRegistrationsByDeviceMethod } from "./api/listRegistrationsByDevice.js";
+import { listRegistrationsByChannel as listRegistrationsByChannelMethod } from "./api/listRegistrationsByChannel.js";
 import { listRegistrationsByTag as listRegistrationsByTagMethod } from "./api/listRegistrationsByTag.js";
 import { listRegistrations as listRegistrationsMethod } from "./api/listRegistrations.js";
 import { scheduleNotification as scheduleNotificationMethod } from "./api/scheduleNotification.js";
@@ -44,7 +44,6 @@ import { sendNotification as sendNotificationMethod } from "./api/sendNotificati
 import { submitNotificationHubJob as submitNotificationHubJobMethod } from "./api/submitNotificationHubJob.js";
 import { updateInstallation as updateInstallationMethod } from "./api/updateInstallation.js";
 import { updateRegistration as updateRegistrationMethod } from "./api/updateRegistration.js";
-import { RegistrationDevice } from "./models/device.js";
 
 /**
  * This represents a client for Azure Notification Hubs to manage installations and send
@@ -207,15 +206,15 @@ export class NotificationHubsClient {
 
   /**
    * Gets all registrations for the notification hub with the given device information and options.
-   * @param device - The device information to query per PNS type.
+   * @param channel - The registration channel information to query per PNS type.
    * @param options - The options for querying the registrations such as $top.
    * @returns A paged async iterable containing all of the registrations for the notification hub.
    */
-  listRegistrationsByDevice(
-    device: RegistrationDevice,
+  listRegistrationsByChannel(
+    channel: RegistrationChannel,
     options: RegistrationQueryLimitOptions = {}
   ): PagedAsyncIterableIterator<RegistrationDescription> {
-    return listRegistrationsByDeviceMethod(this._client, device, options);
+    return listRegistrationsByChannelMethod(this._client, channel, options);
   }
 
   /**